+++ interface/modules/add_items.rb
@@ -1,26 +1,12 @@
 require './classes/book'
-require './classes/label'
 require './classes/music_album'
 require './classes/genre'
 require './classes/author'
 require './classes/game'
+require_relative 'add_category'
 
 module CreateItems
-  def create_label
-    print 'Enter the book label title: '
-    title = gets.chomp
-    print 'Enter the book color: '
-    color = gets.chomp
-    new_label = Label.new(title, color)
-
-    if @labels.one? { |el| el.title == title && el.color == color }
-      return @labels.select { |el| el.title == title && el.color == color }
-    end
-
-    @labels << new_label
-    new_label
-  end
-
+  include CreateCategory
   def add_book
     print 'Enter publish date: '
     publish_date = gets.chomp
@@ -52,17 +38,6 @@ module CreateItems
     cover_state
   end
 
-  def create_genre
-    print 'Enter the album genre: '
-    name = gets.chomp
-    new_genre = Genre.new(name)
-
-    return @genres.select { |genre| genre.name == name } if @genres.one? { |genre| genre.name == name }
-
-    @genres << new_genre
-    new_genre
-  end
-
   def add_music_album
     print 'Enter the music album publish date: '
     publish_date = gets.chomp
@@ -81,20 +56,6 @@ module CreateItems
     new_music_album.genre = genre
     @music_album << new_music_album
     puts 'The music album was created successfully'
-  end
-
-  def create_author
-    print 'Enter the author\'s first name: '
-    first_name = gets.chomp
-    print 'Enter the author\'s last name: '
-    last_name = gets.chomp
-    if @authors.any? { |author| author.first_name == first_name && author.last_name == last_name }
-      puts "The author #{first_name} #{last_name} already exists"
-    else
-      new_author = Author.new(first_name, last_name)
-      @authors << new_author
-    end
-    new_author
   end
 
   def add_game
